# endpoints/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "inSpector Proxy API Rest"
  },
  "host": "inspector-develop.maplander.com",
  "basePath": "/endpoints/v1",
  "tags": [
    {
      "name": "Business Card",
      "description": "Business Card Service"
    },
    {
      "name": "SASISOPA",
      "description": "SASISOPA Service generate attached files"
    },
    {
      "name": "Utils",
      "description": "Global services"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/bc": {
      "get": {
        "tags": [
          "Business Card"
        ],
        "summary": "Create a Business Card",
        "description": "This method creates a first Business Card for any user, compatible with inSpector and Business Card applications, time response between 1-5s",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "required": true,
            "description": "User name"
          },
          {
            "in": "query",
            "name": "lastName",
            "type": "string",
            "required": true,
            "description": "User last name"
          },
          {
            "in": "query",
            "name": "company",
            "type": "string",
            "required": true,
            "description": "User company name"
          },
          {
            "in": "query",
            "name": "phone",
            "type": "string",
            "required": true,
            "description": "User phone (without country code)"
          },
          {
            "in": "query",
            "name": "workPosition",
            "type": "string",
            "required": true,
            "description": "User work position"
          },
          {
            "in": "query",
            "name": "email",
            "type": "string",
            "required": true,
            "description": "User email"
          },
          {
            "in": "query",
            "name": "countryCode",
            "type": "string",
            "required": true,
            "description": "User country code"
          },
          {
            "in": "query",
            "name": "industryCode",
            "type": "integer",
            "required": true,
            "description": "User industry code"
          },
          {
            "in": "query",
            "name": "website",
            "type": "string",
            "description": "User website"
          },
          {
            "in": "query",
            "name": "profileImage",
            "type": "string",
            "description": "User profile blob image"
          },
          {
            "in": "query",
            "name": "profileImageThumbnail",
            "type": "string",
            "description": "User profile image thumbnail"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BCResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BCResponse"
            }
          },
          "422": {
            "description": "Incomplete params",
            "schema": {
              "$ref": "#/definitions/BCResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/BCResponse"
            }
          }
        }
      }
    },
    "/pdf": {
      "get": {
        "tags": [
          "SASISOPA"
        ],
        "summary": "Create a attached file with custom information",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "stationId",
            "type": "string",
            "required": true,
            "description": "Station ID"
          },
          {
            "in": "query",
            "name": "isSGM",
            "type": "boolean",
            "default": false,
            "description": "Is SGM document"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SasisopaResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/SasisopaResponse"
            }
          },
          "422": {
            "description": "Incomplete params",
            "schema": {
              "$ref": "#/definitions/SasisopaResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/SasisopaResponse"
            }
          }
        }
      }
    },
    "/joinPDF": {
      "get": {
        "tags": [
          "SASISOPA"
        ],
        "summary": "Create a PDF file joining multiples files",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/pdf",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "stationId",
            "type": "string",
            "required": true,
            "description": "Station ID"
          },
          {
            "in": "query",
            "name": "isSGM",
            "type": "boolean",
            "default": false,
            "description": "Is SGM document"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "422": {
            "description": "Incomplete params",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    },
    "/download": {
      "get": {
        "tags": [
          "Utils"
        ],
        "summary": "Download file as Buffer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/octet-stream",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "encryptingString",
            "type": "string",
            "required": true,
            "description": "Url File or Image encrypted"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "422": {
            "description": "Incomplete params",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/DefaultResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "DefaultResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "Status code in the response"
        },
        "description": {
          "type": "string",
          "description": "Information on response"
        }
      }
    },
    "BCResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "Status code in the response"
        },
        "description": {
          "type": "string",
          "description": "Information on response"
        },
        "item": {
          "$ref": "#/definitions/UBCLite"
        }
      }
    },
    "SasisopaResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "Status code in the response"
        },
        "description": {
          "type": "string",
          "description": "Information on response"
        },
        "item": {
          "$ref": "#/definitions/FullSasisopa"
        }
      }
    },
    "UBCLite": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Id of Entity BCard in Business Card BakcEnd"
        },
        "userId": {
          "type": "string",
          "description": "Id of Entity User in Business Card BackEnd"
        },
        "cardThumbnail": {
          "type": "string",
          "description": "URL image of Business Card"
        },
        "dynamicLink": {
          "type": "string",
          "description": "Dynamic link for share Business Card"
        },
        "dynamicSignIn": {
          "type": "string",
          "description": "Dynamic link for sign in Business Card"
        }
      }
    },
    "FullSasisopa": {
      "type": "object",
      "properties": {
        "date": {
          "type": "number",
          "description": "Date for file availability"
        }
      }
    }
  }
}